moved universal 'TreeView' to utils
