+++ openpype/tools/ayon_workfiles/widgets/files_widget_published.py
@@ -5,9 +5,10 @@ from openpype.style import (
     get_default_entity_icon_color,
     get_disabled_entity_icon_color,
 )
+from openpype.tools.utils import TreeView
 from openpype.tools.utils.delegates import PrettyTimeDelegate
 
-from .utils import TreeView, BaseOverlayFrame
+from .utils import BaseOverlayFrame
 
 
 REPRE_ID_ROLE = QtCore.Qt.UserRole + 1
